Turn utils.diffStrings into a magicpen style (stringDiff).
