+++ lib/styles.js
@@ -1,4 +1,6 @@
 var utils = require('./utils');
+var stringDiff = require('diff');
+var specialCharRegExp = require('./specialCharRegExp');
 
 module.exports = function (expect) {
     expect.installTheme({ styles: {
@@ -339,5 +341,121 @@ module.exports = function (expect) {
         if (isTopLevel && !pen.isEmpty() && (pen.format === this.format || !pen.format)) {
             this.sp().commentBlock(pen);
         }
+    });
+
+    expect.addStyle('stringDiff', function (actual, expected, options) {
+        options = options || {};
+        var type = options.type || 'WordsWithSpace';
+
+        function addStringToOutput(output, text, baseStyle, specialCharStyle) {
+            if (options.markUpSpecialCharacters) {
+                text.split(specialCharRegExp).forEach(function (part) {
+                    if (specialCharRegExp.test(part)) {
+                        output[specialCharStyle || baseStyle](utils.escapeChar(part));
+                    } else {
+                        output[baseStyle](part);
+                    }
+                });
+            } else {
+                output[baseStyle](text);
+            }
+            return output;
+        }
+
+        var diffLines = [];
+        var lastPart;
+        stringDiff.diffLines(actual, expected).forEach(function (part) {
+            if (lastPart && lastPart.added && part.removed) {
+                diffLines.push({
+                    oldValue: part.value,
+                    newValue: lastPart.value,
+                    replaced: true
+                });
+                lastPart = null;
+            } else if (lastPart) {
+                diffLines.push(lastPart);
+                lastPart = part;
+            } else {
+                lastPart = part;
+            }
+        });
+        if (lastPart) {
+            diffLines.push(lastPart);
+        }
+
+        diffLines.forEach(function (part, index) {
+            var endsWithNewline = /\n$/.test(part.value);
+            var value;
+            if (part.replaced) {
+                var oldLine = this.clone();
+                var newLine = this.clone();
+                var oldValue = part.oldValue;
+                var newValue = part.newValue;
+                var oldEndsWithNewline = oldValue.slice(-1) === '\n';
+                var newEndsWithNewline = newValue.slice(-1) === '\n';
+                if (oldEndsWithNewline) {
+                    oldValue = oldValue.slice(0, -1);
+                }
+                if (newEndsWithNewline) {
+                    newValue = newValue.slice(0, -1);
+                }
+
+                stringDiff['diff' + type](oldValue, newValue).forEach(function (part) {
+                    if (part.added) {
+                        addStringToOutput(newLine, part.value, 'diffAddedHighlight', 'diffAddedSpecialChar');
+                    } else if (part.removed) {
+                        addStringToOutput(oldLine, part.value, 'diffRemovedHighlight', 'diffRemovedSpecialChar');
+                    } else {
+                        newLine.diffAddedLine(part.value);
+                        oldLine.diffRemovedLine(part.value);
+                    }
+                });
+                oldLine.prependLinesWith(this.clone().diffRemovedLine('-'));
+                newLine.prependLinesWith(this.clone().diffAddedLine('+'));
+
+                if (oldEndsWithNewline && !newEndsWithNewline) {
+                    oldLine.diffRemovedSpecialChar('\\n');
+                }
+
+                if (newEndsWithNewline && !oldEndsWithNewline) {
+                    newLine.diffAddedSpecialChar('\\n');
+                }
+
+                this.append(oldLine).nl().append(newLine);
+                if (oldEndsWithNewline && index < diffLines.length - 1) {
+                    this.nl();
+                }
+            } else if (part.added) {
+                value = endsWithNewline ?
+                    part.value.slice(0, -1) :
+                    part.value;
+
+                this.append(function () {
+                    addStringToOutput(this, value, 'diffAddedLine').prependLinesWith(function () {
+                        this.diffAddedLine('+');
+                    });
+                });
+
+                if (endsWithNewline) {
+                    this.nl();
+                }
+            } else if (part.removed) {
+                value = endsWithNewline ?
+                    part.value.slice(0, -1) :
+                    part.value;
+
+                this.append(function () {
+                    addStringToOutput(this, value, 'diffRemovedLine').prependLinesWith(function () {
+                        this.diffRemovedLine('-');
+                    });
+                });
+
+                if (endsWithNewline) {
+                    this.nl();
+                }
+            } else {
+                this.text(part.value.replace(/^(.)/gm, ' $1'));
+            }
+        }, this);
     });
 };
