+++ lib/types.js
@@ -917,7 +917,7 @@ module.exports = function (expect) {
                 diff: output,
                 inline: false
             };
-            utils.diffStrings(String(actual), String(expected), output, {type: 'Chars', markUpSpecialCharacters: true});
+            output.stringDiff(String(actual), String(expected), {type: 'Chars', markUpSpecialCharacters: true});
             return result;
         }
     });
@@ -999,7 +999,7 @@ module.exports = function (expect) {
             if (Math.max(actual.length, expected.length) > this.diffLimit) {
                 result.diff.jsComment('Diff suppressed due to size > ' + this.diffLimit);
             } else {
-                result.diff = utils.diffStrings(this.hexDump(actual), this.hexDump(expected), output, {type: 'Chars', markUpSpecialCharacters: false})
+                result.diff.stringDiff(this.hexDump(actual), this.hexDump(expected), {type: 'Chars', markUpSpecialCharacters: false})
                     .replaceText(/[\x00-\x1f\x7f-\xff␊␍]/g, '.').replaceText(/[│ ]/g, function (styles, content) {
                         this.text(content);
                     });
@@ -1047,7 +1047,7 @@ module.exports = function (expect) {
                 diff: output,
                 inline: false
             };
-            utils.diffStrings(actual, expected, output, {type: 'WordsWithSpace', markUpSpecialCharacters: true});
+            output.stringDiff(actual, expected, {type: 'WordsWithSpace', markUpSpecialCharacters: true});
             return result;
         }
     });
